# perf: cache reserved bind variables in queries