--- a/go/vt/sqlparser/redact_query.go
+++ b/go/vt/sqlparser/redact_query.go
@@ -23,13 +23,13 @@ func RedactSQLQuery(sql string) (string, error) {
 	bv := map[string]*querypb.BindVariable{}
 	sqlStripped, comments := SplitMarginComments(sql)
 
-	stmt, err := Parse(sqlStripped)
+	stmt, reservedVars, err := Parse2(sqlStripped)
 	if err != nil {
 		return "", err
 	}
 
 	prefix := "redacted"
-	err = Normalize(stmt, bv, prefix)
+	err = Normalize(stmt, reservedVars, bv, prefix)
 	if err != nil {
 		return "", err
 	}