--- a/go/vt/vtgate/planbuilder/explain.go
+++ b/go/vt/vtgate/planbuilder/explain.go
@@ -29,13 +29,13 @@ import (
 )
 
 // Builds an explain-plan for the given Primitive
-func buildExplainPlan(stmt sqlparser.Explain, vschema ContextVSchema) (engine.Primitive, error) {
+func buildExplainPlan(stmt sqlparser.Explain, reservedVars sqlparser.BindVars, vschema ContextVSchema) (engine.Primitive, error) {
 	switch explain := stmt.(type) {
 	case *sqlparser.ExplainTab:
 		return explainTabPlan(explain, vschema)
 	case *sqlparser.ExplainStmt:
 		if explain.Type == sqlparser.VitessType {
-			return buildVitessTypePlan(explain, vschema)
+			return buildVitessTypePlan(explain, reservedVars, vschema)
 		}
 		return buildOtherReadAndAdmin(sqlparser.String(explain), vschema)
 	}
@@ -61,8 +61,8 @@ func explainTabPlan(explain *sqlparser.ExplainTab, vschema ContextVSchema) (engi
 	}, nil
 }
 
-func buildVitessTypePlan(explain *sqlparser.ExplainStmt, vschema ContextVSchema) (engine.Primitive, error) {
-	innerInstruction, err := createInstructionFor(sqlparser.String(explain.Statement), explain.Statement, vschema)
+func buildVitessTypePlan(explain *sqlparser.ExplainStmt, reservedVars sqlparser.BindVars, vschema ContextVSchema) (engine.Primitive, error) {
+	innerInstruction, err := createInstructionFor(sqlparser.String(explain.Statement), explain.Statement, reservedVars, vschema)
 	if err != nil {
 		return nil, err
 	}