--- a/go/vt/vtgate/planbuilder/fallback_planner.go
+++ b/go/vt/vtgate/planbuilder/fallback_planner.go
@@ -29,28 +29,28 @@ type fallbackPlanner struct {
 
 var _ selectPlanner = (*fallbackPlanner)(nil).plan
 
-func (fp *fallbackPlanner) safePrimary(query string) func(sqlparser.Statement, ContextVSchema) (engine.Primitive, error) {
+func (fp *fallbackPlanner) safePrimary(query string) func(sqlparser.Statement, sqlparser.BindVars, ContextVSchema) (engine.Primitive, error) {
 	primaryF := fp.primary(query)
-	return func(stmt sqlparser.Statement, vschema ContextVSchema) (res engine.Primitive, err error) {
+	return func(stmt sqlparser.Statement, reservedVars sqlparser.BindVars, vschema ContextVSchema) (res engine.Primitive, err error) {
 		defer func() {
 			// if the primary planner panics, we want to catch it here so we can fall back
 			if r := recover(); r != nil {
 				err = fmt.Errorf("%v", r) // not using vterror since this will only be used for logging
 			}
 		}()
-		res, err = primaryF(stmt, vschema)
+		res, err = primaryF(stmt, reservedVars, vschema)
 		return
 	}
 }
 
-func (fp *fallbackPlanner) plan(query string) func(sqlparser.Statement, ContextVSchema) (engine.Primitive, error) {
+func (fp *fallbackPlanner) plan(query string) func(sqlparser.Statement, sqlparser.BindVars, ContextVSchema) (engine.Primitive, error) {
 	primaryF := fp.safePrimary(query)
 	backupF := fp.fallback(query)
 
-	return func(stmt sqlparser.Statement, vschema ContextVSchema) (engine.Primitive, error) {
-		res, err := primaryF(stmt, vschema)
+	return func(stmt sqlparser.Statement, reservedVars sqlparser.BindVars, vschema ContextVSchema) (engine.Primitive, error) {
+		res, err := primaryF(stmt, reservedVars, vschema)
 		if err != nil {
-			return backupF(stmt, vschema)
+			return backupF(stmt, reservedVars, vschema)
 		}
 		return res, nil
 	}