… go/test/fuzzing/parser_fuzzer.go
… go/vt/sqlparser/normalizer_test.go
… go/vt/sqlparser/sql.go
… go/vt/vtgate/planbuilder/fallback_planner_test.go
… go/vt/vtgate/planbuilder/from.go
… go/vt/vtgate/planbuilder/insert.go
… go/vt/vtgate/planbuilder/join.go
… go/vt/vtgate/planbuilder/locktables.go
… go/vt/vtgate/planbuilder/route_planning.go
… go/vt/vtgate/planbuilder/select.go
… go/vt/vtgate/planbuilder/union.go
… go/vt/vtgate/planbuilder/update.go