@vmg 2021-03-17
@harshit-gangal: that was the first design I attempted:

```go
type BindableStatement struct {
    Statement
    KnownBinds BindVars
}
```

I tried wrapping all the statements that had bindable variables with a `Statement` type that caches them, but this is extremely complicated to wire up because the codebase is littered with hundreds of `(type)` checks on statement types that simply "break" its logic when we introduce this new wrapping statement but do _not_ fail to compile. I tried to fix all the breakage for a couple hours, but it was pointless. After that, I switched to explicitly wiring up the bind variables, because changing function signatures makes the compiler guide you to the places that need changing.